Criado modal de criacao de escala e refatorado nomeclatura de componentes
