+++ src/ui/agenda/Agenda.view.jsx
