+++ src/ui/common/BottomMenuBar.component.jsx
@@ -1,16 +1,40 @@
 import * as React from 'react';
-import {StyleSheet} from 'react-native';
+import {StyleSheet, Keyboard} from 'react-native';
 import {Appbar, FAB, useTheme} from 'react-native-paper';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 
 const BOTTOM_APPBAR_HEIGHT = 80;
 const MEDIUM_FAB_HEIGHT = 56;
 
-const AgendaView = ({navigation}) => {
+const BottomMenuBar = ({navigation, onPressFAB}) => {
   const {bottom} = useSafeAreaInsets();
   const theme = useTheme();
 
-  return (
+  const [isKeyboardVisible, setKeyboardVisible] = React.useState(false);
+
+  React.useEffect(() => {
+    const keyboardDidShowListener = Keyboard.addListener(
+      'keyboardDidShow',
+      () => {
+        setKeyboardVisible(true); // or some other action
+      },
+    );
+    const keyboardDidHideListener = Keyboard.addListener(
+      'keyboardDidHide',
+      () => {
+        setKeyboardVisible(false); // or some other action
+      },
+    );
+
+    return () => {
+      keyboardDidHideListener.remove();
+      keyboardDidShowListener.remove();
+    };
+  }, []);
+
+  return isKeyboardVisible ? (
+    <></>
+  ) : (
     <Appbar
       style={[
         styles.bottom,
@@ -32,7 +56,7 @@ const AgendaView = ({navigation}) => {
         mode="flat"
         size="medium"
         icon="plus"
-        onPress={()=>{}}
+        onPress={onPressFAB}
         style={[
           styles.fab,
           {top: (BOTTOM_APPBAR_HEIGHT - MEDIUM_FAB_HEIGHT) / 2},
@@ -56,4 +80,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default AgendaView;
+export default BottomMenuBar;
