+++ src/ui/equipe/Equipe.view.jsx
